+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
-## [UNREALESED] - 2022-9-3
+## [2022.3.1] - 2022-15-3
 
 ### Added
 
@@ -25,7 +25,7 @@ and this project adheres to [CalVer](https://calver.org/).
   and absolute `px. `height` can only be absolute:
 
 ```json
-  {
+{
   "vectorTable": {
     "position": "right",
     "width": "30%",